# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EdgeX Ops Intelligence - Liquidity API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/readiness": {
      "get": {
        "responses": {
          "200": {
            "description": "ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessResponse"
                }
              }
            }
          },
          "503": {
            "description": "not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/snapshot/liquidity": {
      "get": {
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "liquidity tab",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiquiditySnapshot"
                }
              }
            }
          }
        }
      }
    },
    "/api/snapshot/quality": {
      "get": {
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "quality tab",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualitySnapshot"
                }
              }
            }
          }
        }
      }
    },
    "/api/snapshot/share": {
      "get": {
        "parameters": [
          {
            "name": "window",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "24h",
                "7d",
                "30d"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "market share tab",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareSnapshot"
                }
              }
            }
          }
        }
      }
    },
    "/api/snapshot/top30": {
      "get": {
        "parameters": [
          {
            "name": "surface",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "platform",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "top30 tab",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Top30Snapshot"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiStatus": {
        "type": "string",
        "enum": [
          "complete",
          "partial",
          "aggregated_orderbook",
          "ws_limited_depth",
          "stale",
          "unsupported",
          "error",
          "insufficient_history"
        ]
      },
      "DataFreshness": {
        "type": "string",
        "enum": [
          "live",
          "delayed"
        ]
      },
      "NumberMap": {
        "type": "object",
        "additionalProperties": {
          "type": "number"
        }
      },
      "StringMap": {
        "type": "object",
        "additionalProperties": {
          "type": "string"
        }
      },
...
      "ReadinessResponse": {
        "type": "object",
        "required": [
          "ready",
          "build_version",
          "checks"
        ],
        "properties": {
          "ready": {
            "type": "boolean"
          },
          "build_version": {
            "type": "string"
          },
          "checks": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
      "DepthTierMetrics": {
        "type": "object",
        "required": [
          "bid_usd",
          "ask_usd",
          "total_usd"
        ],
        "properties": {
          "bid_usd": {
            "type": "number"
          },
          "ask_usd": {
            "type": "number"
          },
          "total_usd": {
            "type": "number"
          },
          "depth_status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "partial_reason": {
            "type": "string"
          },
          "depth_source": {
            "type": "string"
          },
          "source_id": {
            "type": "string"
          },
          "source_endpoint": {
            "type": "string"
          },
          "levels_returned": {
            "type": "integer"
          },
          "bid_levels_returned": {
            "type": "integer"
          },
          "ask_levels_returned": {
            "type": "integer"
          },
          "api_level_cap": {
            "type": "integer"
          },
          "farthest_bid_pct": {
            "type": "number"
          },
          "farthest_ask_pct": {
            "type": "number"
          },
          "farthest_distance_pct": {
            "type": "number"
          },
          "aggregation_params": {
            "$ref": "#/components/schemas/StringMap"
          },
          "strict_complete": {
            "type": "boolean"
          },
          "display_available": {
            "type": "boolean"
          },
          "policy_acceptance": {
            "type": "string"
          },
          "physical_limit": {
            "type": "boolean"
          },
          "unofficial_ui_endpoint": {
            "type": "boolean"
          }
        }
      },
      "PlatformRow": {
        "type": "object",
        "required": [
          "platform",
          "display_symbol",
          "snapshot_ts",
          "source_endpoint",
          "depth_status",
          "depth_by_tier",
          "buy_slippage_bp",
          "sell_slippage_bp"
        ],
        "properties": {
          "platform": {
            "type": "string"
          },
          "display_symbol": {
            "type": "string"
          },
          "snapshot_ts": {
            "type": "string",
            "format": "date-time"
          },
          "source_endpoint": {
            "type": "string"
          },
          "depth_status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "partial_reason": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "data_freshness": {
            "$ref": "#/components/schemas/DataFreshness"
          },
          "last_collection_status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "last_collection_error": {
            "type": "string"
          },
          "last_collection_ts": {
            "type": "string",
            "format": "date-time"
          },
          "mid_price": {
            "type": "number"
          },
          "spread_bp": {
            "type": "number"
          },
          "imbalance_pct": {
            "type": "number"
          },
          "depth_by_tier": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/DepthTierMetrics"
            }
          },
          "vs_median_by_tier": {
            "$ref": "#/components/schemas/NumberMap"
          },
          "rank_0_1": {
            "type": "integer"
          },
          "depth_status_label": {
            "type": "string"
          },
          "buy_slippage_bp": {
            "$ref": "#/components/schemas/NumberMap"
          },
          "sell_slippage_bp": {
            "$ref": "#/components/schemas/NumberMap"
          },
          "worst_slippage_bp": {
            "$ref": "#/components/schemas/NumberMap"
          },
          "vs_median_spread_bp": {
            "type": "number",
            "nullable": true
          },
          "vs_median_slippage_bp": {
            "$ref": "#/components/schemas/NumberMap"
          },
          "verdict": {
            "type": "string"
          },
          "funding": {
            "$ref": "#/components/schemas/PlatformFundingRate"
          }
        }
      },
      "PlatformFundingRate": {
        "type": "object",
        "required": [
          "platform",
          "status"
        ],
        "properties": {
          "platform": {
            "type": "string"
          },
          "display_symbol": {
            "type": "string"
          },
          "rate_8h": {
            "type": "number",
            "nullable": true
          },
          "rate_native": {
            "type": "number",
            "nullable": true
          },
          "period_hours": {
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "source": {
            "type": "string"
          },
          "source_endpoint": {
            "type": "string"
          },
          "snapshot_ts": {
            "type": "string",
            "format": "date-time"
          },
          "vs_median_8h": {
            "type": "number",
            "nullable": true
          },
          "rank_positive": {
            "type": "integer",
            "nullable": true
          },
          "rank_negative": {
            "type": "integer",
            "nullable": true
          }
        }
      },
      "LiquidityKPIs": {
        "type": "object",
        "properties": {
          "edgex_depth_by_tier": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/DepthTierMetrics"
            }
          },
          "edgex_vs_median_by_tier": {
            "$ref": "#/components/schemas/NumberMap"
          },
          "strict_competitor_median_by_tier": {
            "$ref": "#/components/schemas/NumberMap"
          },
          "competitor_median_by_tier": {
            "$ref": "#/components/schemas/NumberMap"
          },
          "edgex_spread_bp": {
            "type": "number"
          },
          "edgex_spread_10m_bp": {
            "type": "number"
          },
          "edgex_spread_10m_status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "edgex_24h_share_pct": {
            "type": "number"
          },
          "edgex_24h_share_status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "symbol_share_7d_pct": {
            "type": "number"
          },
          "symbol_share_7d_status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "symbol_share_wow_status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "edgex_funding_rate_8h": {
            "type": "number",
            "nullable": true
          },
          "edgex_funding_rate_8h_status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "edgex_funding_rate_period_hours": {
            "type": "integer"
          },
          "edgex_funding_rate_vs_median_8h": {
            "type": "number",
            "nullable": true
          },
          "competitor_funding_rate_median_8h": {
            "type": "number",
            "nullable": true
          },
          "competitor_funding_rate_median_8h_status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "competitor_funding_rate_median_8h_samples": {
            "type": "integer"
          }
        }
      },
      "QualityKPIs": {
        "type": "object",
        "properties": {
          "edgex_funding_rate_8h": {
            "type": "number",
            "nullable": true
          },
          "edgex_funding_rate_8h_status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "edgex_funding_rate_period_hours": {
            "type": "integer"
          },
          "competitor_funding_rate_median_8h": {
            "type": "number",
            "nullable": true
          },
          "competitor_funding_rate_median_8h_status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "competitor_funding_rate_median_8h_samples": {
            "type": "integer"
          }
        }
      },
      "LiquiditySnapshot": {
        "type": "object",
        "required": [
          "symbol",
          "snapshot_ts",
          "rows"
        ],
        "properties": {
          "symbol": {
            "type": "string"
          },
          "snapshot_ts": {
            "type": "string",
            "format": "date-time"
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlatformRow"
            }
          },
          "competitor_median_by_tier": {
            "$ref": "#/components/schemas/NumberMap"
          },
          "strict_competitor_median_by_tier": {
            "$ref": "#/components/schemas/NumberMap"
          },
          "kpis": {
            "$ref": "#/components/schemas/LiquidityKPIs"
          }
        }
      },
      "QualitySnapshot": {
        "type": "object",
        "required": [
          "symbol",
          "snapshot_ts",
          "slippage_buckets_usd",
          "rows"
        ],
        "properties": {
          "symbol": {
            "type": "string"
          },
          "snapshot_ts": {
            "type": "string",
            "format": "date-time"
          },
          "slippage_buckets_usd": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlatformRow"
            }
          },
          "kpis": {
            "$ref": "#/components/schemas/QualityKPIs"
          }
        }
      },
      "ShareRow": {
        "type": "object",
        "required": [
          "platform"
        ],
        "properties": {
          "rank": {
            "type": "integer"
          },
          "platform": {
            "type": "string"
          },
          "raw_volume_usd": {
            "type": "number"
          },
          "adjusted_volume_usd": {
            "type": "number"
          },
          "adjusted_volume_24h_usd": {
            "type": "number"
          },
          "adjusted_volume_total_usd": {
            "type": "number"
          },
          "share_pct": {
            "type": "number"
          },
          "denominator_pct": {
            "type": "number"
          },
          "discount": {
            "type": "number"
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "data_source": {
            "type": "string"
          },
          "days_seen": {
            "type": "integer"
          },
          "days_window": {
            "type": "integer"
          }
        }
      },
      "ShareTrend": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "points": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": {
                "oneOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "string"
                  }
                ]
              }
            }
          }
        }
      },
      "ShareHistory": {
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/ApiStatus"
        }
      },
      "ShareKPIs": {
        "type": "object",
        "properties": {
          "edgex_share_pct": {
            "type": "number"
          },
          "edgex_total_volume_usd": {
            "type": "number"
          },
          "denominator_usd": {
            "type": "number"
          }
        }
      },
      "ShareSnapshot": {
        "type": "object",
        "required": [
          "window",
          "snapshot_ts",
          "rows"
        ],
        "properties": {
          "window": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "reason": {
            "type": "string"
          },
          "snapshot_ts": {
            "type": "string",
            "format": "date-time"
          },
          "denominator_usd": {
            "type": "number"
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShareRow"
            }
          },
          "kpis": {
            "$ref": "#/components/schemas/ShareKPIs"
          },
          "history": {
            "$ref": "#/components/schemas/ShareHistory"
          },
          "trend": {
            "$ref": "#/components/schemas/ShareTrend"
          }
        }
      },
      "Top30Row": {
        "type": "object",
        "required": [
          "rank",
          "platform",
          "symbol"
        ],
        "properties": {
          "rank": {
            "type": "integer"
          },
          "platform": {
            "type": "string"
          },
          "symbol": {
            "type": "string"
          },
          "volume_24h_usd": {
            "type": "number"
          },
          "volume_7d_usd": {
            "type": "number"
          },
          "delta_7d_pct": {
            "type": "number"
          },
          "volume_7d_status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "delta_7d_status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "edgex_listed": {
            "type": "boolean"
          },
          "edgex_listed_status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "competitor_top30_coverage": {
            "type": "integer"
          },
          "competitor_top30_coverage_status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "suggested_action": {
            "type": "string"
          },
          "suggested_action_status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "data_source": {
            "type": "string"
          },
          "source_endpoint": {
            "type": "string"
          },
          "snapshot_ts": {
            "type": "string",
            "format": "date-time"
          },
          "error": {
            "type": "string"
          }
        }
      },
      "Top30Snapshot": {
        "type": "object",
        "required": [
          "surface",
          "platform",
          "snapshot_ts",
          "rows"
        ],
        "properties": {
          "surface": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "snapshot_ts": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Top30Row"
            }
          }
        }
      },
...
    }
  }
}